pyproject.toml
[project]
name = "steam-content-based-recsys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.13.0",
    "nltk>=3.9.2",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
]
